Exercise: React State
In this exercise we went over set state and use state in React. We set the typed search term to a state variable then displayed it in the search component. When the user types, set state is called. Once set state is called React re-renders the search component which displays the new search state.