Exercise: React Component Composition
In this exercise we removed the label property from the inputwithlabel component and instead used the inputwithlabel component in the app component as an HTML element. Inside the inputwithlabel brackets we put our label text with a strong HTML element and accessed the HTML contents via the children property in the search component to set the label properly.