Exercise: React Conditional Rendering
In this exercise we added some JSX that checks a isLoading state parameter to see if we need to render a Loading <p> tag or if we can render the game list component. Given server lag this conditional rendering is needed so a user understands the data is loading. We also added another conditional render that shows an error <p> tag when the server has an issue.