Exercise: React Impossible States
In this exercise we went over impossible states / error states. We simulated an error in rendering the list of games to see how the error <p> was displayed. The error left the loading <p> tag showing indefinitely so we consolidated the state into one object and relegated the changing of the state to the gameReducer.