Exercise: Explicit Data Fetching with React
In this exercise we changed the typeahead to a search where the user has to click a button to submit a search. I had to take some time out to make an API because a lot of the exercises relied on an API where the filtering does not happen on the client side so I made one.